+++ packages/iceworks-server/src/lib/adapter/modules/page/index.ts
@@ -9,7 +9,7 @@ import * as _ from 'lodash';
 import * as mv from 'mv';
 import * as mkdirp from 'mkdirp';
 import * as upperCamelCase from 'uppercamelcase';
-import * as kebabCase from 'kebab-case';
+import * as uniqBy from 'lodash.uniqby';
 import { getAndExtractTarball } from 'ice-npm-utils';
 import scanDirectory from '../../../scanDirectory';
 import getNpmClient from '../../../getNpmClient';
@@ -25,9 +25,7 @@ const readFileAsync = util.promisify(fs.readFile);
 const lstatAsync = util.promisify(fs.lstat);
 const mvAsync = util.promisify(mv);
 
-const loadTemplate = async () => {
-  const fileName = 'template.jsx';
-  const filePath = path.join(__dirname, `${fileName}.ejs`);
+const loadTemplate = async (fileName: string, filePath: string) => {
   const fileStr = await readFileAsync(filePath, 'utf-8');
   const compile = ejs.compile(fileStr);
   return {
@@ -46,11 +44,20 @@ export default class Page implements IPageModule {
 
   private readonly componentDirName: string = 'components';
 
-  constructor(params: {project: IProject; storage: any }) {
+  public readonly templateFileName: string;
+
+  public readonly templateFilePath: string;
+
+  public readonly prettierParseType: string;
+
+  constructor(params: { project: IProject; storage: any }) {
     const { project, storage } = params;
     this.project = project;
     this.storage = storage;
     this.path = path.join(this.project.path, 'src', 'pages');
+    this.templateFileName = 'template.jsx';
+    this.templateFilePath = path.join(__dirname, `${this.templateFileName}.ejs`);
+    this.prettierParseType = 'babel';
   }
 
   private async scanPages(dirPath: string): Promise<IPage[]> {
@@ -120,7 +127,7 @@ export default class Page implements IPageModule {
     await mkdirpAsync(componentsDir);
 
     const iceVersion: string = getIceVersion(projectPackageJSON);
-    const blockName: string = block.alias || upperCamelCase(block.name);
+    const blockName: string = this.generateBlockName(block);
 
     let tarballURL: string;
     try {
@@ -145,8 +152,15 @@ export default class Page implements IPageModule {
       throw error;
     }
 
-    await mkdirpAsync(blockDir);
-    await mvAsync(path.join(blockTempDir, 'src'), blockDir);
+    await mvAsync(path.join(blockTempDir, 'src'), blockDir, {clobber: false});
+  }
+
+  private generateBlockName(block: {name: string}): string {
+    return upperCamelCase(block.name);
+  }
+
+  private checkBlocksName(blocks: {name: string}[]): boolean {
+    return uniqBy(blocks.map((block) => ({ name: this.generateBlockName(block) })), 'name').length !== blocks.length;
   }
 
   public async getAll(): Promise<IPage[]> {
@@ -158,13 +172,15 @@ export default class Page implements IPageModule {
     const { name, blocks } = page;
     const { socket, i18n } = ctx;
 
+    if (this.checkBlocksName(blocks)) {
+      throw new Error(i18n.format('baseAdapter.page.blocks.exist'));
+    }
+
     // create page dir
     socket.emit('adapter.page.create.status', { text: i18n.format('baseAdapter.page.create.createMenu'), percent: 10 });
-    const pageFolderName = upperCamelCase(name);
-    const pageDir = path.join(this.path, pageFolderName);
+    const pageName = upperCamelCase(name);
+    const pageDir = path.join(this.path, pageName);
     await mkdirpAsync(pageDir);
-
-    const pageName = kebabCase(pageFolderName).replace(/^-/, '');
 
     if (fs.readdirSync(pageDir).length > 0) {
       const error: any = new Error(`${name} 页面已存在，不允许覆盖。`);
@@ -182,19 +198,19 @@ export default class Page implements IPageModule {
 
     // create page file
     socket.emit('adapter.page.create.status', { text: i18n.format('baseAdapter.page.create.createFile'), percent: 90 });
-    const template = await loadTemplate();
+   
+    const template = await loadTemplate(this.templateFileName, this.templateFilePath);
     const fileContent = template.compile({
       blocks: blocks.map((block) => {
-        const blockFolderName = block.alias || upperCamelCase(block.name);
-        const blockClassName = upperCamelCase(block.alias || block.name);
+        const blockName = this.generateBlockName(block);
 
         return {
           ...block,
-          className: blockClassName,
-          relativePath: `./${this.componentDirName}/${blockFolderName}`,
+          className: blockName,
+          relativePath: `./${this.componentDirName}/${blockName}`,
         };
       }),
-      className: pageFolderName,
+      className: pageName,
       pageName,
     });
     const fileName = template.fileName
@@ -203,11 +219,11 @@ export default class Page implements IPageModule {
     const dist = path.join(pageDir, fileName);
     const rendered = prettier.format(
       fileContent,
-      { singleQuote: true, trailingComma: 'es5', parser: 'babel' }
+      { singleQuote: true, trailingComma: 'es5', parser: this.prettierParseType }
     );
 
     await writeFileAsync(dist, rendered, 'utf-8');
-    return pageFolderName;
+    return pageName;
   }
 
   public async delete(params: {name: string}): Promise<any> {
@@ -230,6 +246,13 @@ export default class Page implements IPageModule {
 
   public async addBlocks(params: {blocks: IMaterialBlock[]; name?: string }, ctx: IContext): Promise<void> {
     const {blocks, name} = params;
+    const {i18n } = ctx;
+
+    const existBlocks = await this.getBlocks(name);
+    if (this.checkBlocksName(existBlocks.concat(blocks))) {
+      throw new Error(i18n.format('baseAdapter.page.blocks.exist'));
+    }
+    
     await this.downloadBlocksToPage(blocks, name, ctx);
   }